+++ app/src/main/java/com/studhub/app/presentation/listing/add/CreateListingViewModel.kt
@@ -1,4 +1,4 @@
-package com.studhub.app.presentation.ui.listing
+package com.studhub.app.presentation.listing.add
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
