+++ app/src/main/java/com/studhub/app/presentation/listing/add/FakeListingRepository.kt
@@ -1,4 +1,4 @@
-package com.studhub.app.presentation.ui.listing
+package com.studhub.app.presentation.listing.add
 
 import com.studhub.app.core.utils.ApiResponse
 import com.studhub.app.domain.model.Category
