+++ app/src/main/java/com/studhub/app/presentation/listing/browse/BrowseScreen.kt
@@ -1,4 +1,4 @@
-package com.studhub.app.presentation.ui.browse
+package com.studhub.app.presentation.listing.browse
 
 import BrowseContent
 import androidx.compose.foundation.layout.Box
