+++ app/src/main/java/com/studhub/app/presentation/listing/browse/DetailedListingViewModel.kt
@@ -1,9 +1,10 @@
-package com.studhub.app.presentation.listing.add
+package com.studhub.app.presentation.listing.browse
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.studhub.app.core.utils.ApiResponse
 import com.studhub.app.domain.model.Listing
+import com.studhub.app.presentation.listing.add.FakeListingRepository
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
