+++ app/src/main/java/com/studhub/app/presentation/listing/browse/components/PriceChip.kt
@@ -1,10 +1,9 @@
-package com.studhub.app.presentation.ui.browse.components
+package com.studhub.app.presentation.listing.browse.components
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.shape.RoundedCornerShape
