+++ app/src/main/java/com/studhub/app/presentation/ui/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.studhub.app.ui.theme
+package com.studhub.app.presentation.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Shapes
